--- backend/pkg/util/connection_url_builder.go
@@ -1,29 +1,0 @@
-package util
-
-import (
-	"fmt"
-	"os"
-)
-
-// ConnectionURLBuilder func for building URL connection.
-func ConnectionURLBuilder(n string) (string, error) {
-	// Define URL to connection.
-	var url string
-
-	// Switch given names.
-	switch n {
-	case "server":
-		// URL for server connection.
-		url = fmt.Sprintf(
-			"%s:%s",
-			os.Getenv("SERVER_HOST"),
-			os.Getenv("SERVER_PORT"),
-		)
-	default:
-		// Return error message.
-		return "", fmt.Errorf("connection name '%v' is not supported", n)
-	}
-
-	// Return connection URL.
-	return url, nil
-}
